Move Exception Handling to HTTP_STATUS
It's moved to an ControllerAdvice class to be able to use it in the population endpoint as well.
